Extracted SetDateInterval component into its own file
